pyproject.toml
[tool.poetry]
name = "event-agency"
version = "0.1.0"
description = ""
authors = ["mikeyemail.com"]
readme = "README.md"
packages = [{include = "event_agency"}]

[tool.poetry.dependencies]
python = "3.11.9"
django = "^5.0.6"
djangorestframework = "^3.15.1"
celery = "^5.4.0"
redis = "^5.0.4"
django-environ = "^0.11.2"
django-rest-auth = "^0.9.5"
psycopg2-binary = "^2.9.9"
django-filter = "^24.2"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"
xenon = "^0.9.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# === Black ===


# === Isort ===
